Prevent failure when h5pyd returns a group item as None
There appears to be a bug in h5pyd, which prevents loading externally
linked data, because the relative file path of the linked file is not
converted to a valid domain name. A call to a group’s items() function
will return a value of None, which needs to be trapped before
attempting to read other attributes, etc.